requirements.txt
streamlit==1.0.0
pandas==1.3.3
Pillow==8.4.0
